Use of ptrdiff_t.
Rev: src/interpret.h:1.51

+++ src/interpret.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: interpret.h,v 1.50 2000/07/07 03:01:24 hubbe Exp $
+ * $Id: interpret.h,v 1.51 2000/07/07 15:23:56 grubba Exp $
  */
 #ifndef INTERPRET_H
 #define INTERPRET_H
@@ -95,7 +95,7 @@ struct pike_frame
 #define pop_stack() do{ free_svalue(--Pike_sp); debug_check_stack(); }while(0)
 
 #define pop_n_elems(X)						\
- do { int x_=(X); if(x_) { 					\
+ do { ptrdiff_t x_=(X); if(x_) { 					\
    check__positive(x_,("Popping negative number of args.... (%d) \n",x_)); \
    Pike_sp-=x_; debug_check_stack();					\
   free_svalues(Pike_sp,x_,BIT_MIXED);				\
